Alter delete_media_files_attached_to_article() function to delete files only if directory exists

+++ apps/articles/services.py
@@ -215,6 +215,7 @@ def save_media_file_attached_to_article(file: IO, article_id: int) -> tuple[str,
 
 def delete_media_files_attached_to_article(article: Article) -> None:
     article_dir = os.path.join("articles", "uploads", article.author.username, str(article.id))
-    for file in default_storage.listdir(article_dir)[1]:
-        default_storage.delete(os.path.join(article_dir, file))
-    default_storage.delete(article_dir)
+    if default_storage.exists(article_dir):
+        for file in default_storage.listdir(article_dir)[1]:
+            default_storage.delete(os.path.join(article_dir, file))
+        default_storage.delete(article_dir)
